requirements.txt
opencv-python==4.10.0.84
gradio
pillow
boto3
moviepy
